All: Small clean up

+++ src/main/kotlin/com/chattriggers/ctjs/minecraft/objects/display/DisplayLine.kt
@@ -10,8 +10,8 @@ import com.chattriggers.ctjs.triggers.TriggerType
 import com.chattriggers.ctjs.utils.kotlin.External
 import com.chattriggers.ctjs.utils.kotlin.NotAbstract
 import org.lwjgl.input.Mouse
+import org.lwjgl.util.vector.Vector2f
 import org.mozilla.javascript.NativeObject
-import javax.vecmath.Vector2d
 
 @External
 @NotAbstract
@@ -29,7 +29,7 @@ abstract class DisplayLine {
     private var onDragged: OnTrigger? = null
 
     private var mouseState = HashMap<Int, Boolean>()
-    private var draggedState = HashMap<Int, Vector2d>()
+    private var draggedState = HashMap<Int, Vector2f>()
 
     constructor(text: String) {
         setText(text)
@@ -130,7 +130,7 @@ abstract class DisplayLine {
                 handleClicked(button)
                 mouseState[button] = Mouse.isButtonDown(button)
                 if (Mouse.isButtonDown(button))
-                    draggedState[button] = Vector2d(Client.getMouseX().toDouble(), Client.getMouseY().toDouble())
+                    draggedState[button] = Vector2f(Client.getMouseX(), Client.getMouseY())
             }
         }
 
@@ -171,7 +171,7 @@ abstract class DisplayLine {
             button
         ))
 
-        draggedState[button] = Vector2d(Client.getMouseX().toDouble(), Client.getMouseY().toDouble())
+        draggedState[button] = Vector2f(Client.getMouseX(), Client.getMouseY())
     }
 
     private fun drawFullBG(
